requirements.txt
# ----------------------------------------------------------------
# Common Dependencies
# ----------------------------------------------------------------
psycopg2-binary==2.9.7          # PostgreSQL adapter

# ----------------------------------------------------------------
# Database Module Dependencies
# ----------------------------------------------------------------
faker==19.6.2                   # Generate fake data for testing

# ----------------------------------------------------------------
# AI Agent Module Dependencies  
# ----------------------------------------------------------------
langgraph==0.2.28               # LangGraph orchestration framework
langchain==0.3.1                # LangChain core
langchain-community==0.3.1      # LangChain community integrations
python-dotenv==1.0.0            # Environment variable management
sqlparse==0.4.4                 # SQL parsing and validation
tabulate==0.9.0                 # Pretty-print tabular data
